Add ability to tune scheduler subset size
This setting will have the scheduler pic randomly from up to N computes
that pass through the filters. On a busy system or with multiple
schedulers at play this can more evenly distribute the load of instances
across systems.

Leave the setting at 1, the default, which should have no impact. This
will allow us to tune clusters based on needs.

Change-Id: I97fb3eea1b11755c3fbade7910650c7218c482e3