Use correct domain auth args for os_ modules
Our use of OS_DOMAIN_ID exposed a bug in newer openstack-client-config,
and we can work around this bug by using the proper arguments for the
domain details. Using OS_DOMAIN_ID relied on some magic to interpret
this into setting the same value for both a client domain and a project
domain (because they could be different). Doing it the right way avoids
the (broken) magic.

Change-Id: Ie0ec3ec33fc2a34e3fd0e3924885094665c2d6f1