Support a repo based install for oidc
Blue Box is no longer going to file mirror the oidc packages, therefor
we need to support a repo method where the package has been introduced
into an apt repo. There is no public repo for this at this time,
therefor the file method remains the default, albiet back to github for
the file URL. A site with a proper repo can change the method and
override the repository values to get the package from a repo.

Change-Id: I2f09444abbe005f7e7e235515eb134078355e8d9