Bump qemu version to 1.28
In our 3.0.1 release, qemu versions were not capped, and toward the end
of 3.0.1 installs, a newer qemu build was released. Our 3.1.0 release
was pinned to a previous build, which presented a downgrade scenario. To
resolve this, bump the version we pin to.

Change-Id: I53e6193ee6ec51a96229fee90a57afac358f3a3a
(cherry picked from commit dd7515a3b8301432f9e092535ebcd1d4cd537d1d)