Upgrade percona-xtrabackup along with server
Turns out the server requires a version of xtrabackup and it may not
properly express the verison requirement. So we want to upgrade it to
the latest at the same time as the server so the restart works.

Also, this play only runs on one host at a time, no need for run_once
and delegate shenanigans.

Change-Id: I840cc7fd78920f8deb06e59a0b50b04a9979a3cb